Change PID function signatures to use int16_t

+++ Buck_Converter.X/pid.h
@@ -19,8 +19,8 @@ typedef struct {
 } PID_t;
 
 // Func Prototypes
-void PID_Init(PID_t * pid, float Kp, float Ki, float Kd, float out_min, float out_max, float dt);
-float PID_Compute(PID_t * pid, float setpoint, float measured);
+void PID_Init(PID_t * pid, int16_t Kp, int16_t Ki, int16_t Kd, int16_t out_min, int16_t out_max, int16_t dt);
+int16_t PID_Compute(PID_t * pid, int16_t setpoint, int16_t measured);
 void PID_Reset(PID_t *pid);
 
 #endif // PID_H
